Got rid of BaseImageRequestBuilder interface
